+++ packages/components/src/components/Select/Select.stories.tsx
@@ -4,17 +4,20 @@ import { useBoolean } from '@koobiq/react-core';
 import {
   IconAnomaly16,
   IconBug16,
+  IconCrosshairs16,
   IconDesktop16,
   IconServer16,
   IconSwords16,
 } from '@koobiq/react-icons';
 import type { Meta, StoryObj } from '@storybook/react';
 
+import { useListData } from '../../index';
+import type { Selection } from '../../types';
 import { Button } from '../Button';
 import { FlexBox } from '../FlexBox';
 import { Typography } from '../Typography';
 
-import type { SelectProps, SelectKey } from './index.js';
+import type { SelectProps } from './index.js';
 import { Select } from './index.js';
 
 const meta = {
@@ -26,14 +29,15 @@ const meta = {
   subcomponents: {
     'Select.Item': Select.Item,
     'Select.Section': Select.Section,
+    'Select.Divider': Select.Divider,
     'Select.ItemText': Select.ItemText,
   },
   argTypes: {},
-  tags: ['status:deprecated'],
+  tags: ['status:updated'],
 } satisfies Meta<typeof Select>;
 
 export default meta;
-type Story = StoryObj<SelectProps<unknown>>;
+type Story = StoryObj<SelectProps<object>>;
 
 const options = [
   { id: 1, name: 'Bruteforce' },
@@ -54,8 +58,9 @@ export const Base: Story = {
     <Select
       label="Attack type"
       style={{ inlineSize: 200 }}
-      defaultSelectedKey="bruteforce"
       placeholder="Select an option"
+      selectionMode="multiple"
+      isClearable
       {...args}
     >
       <Select.Item key="bruteforce">Bruteforce</Select.Item>
@@ -103,7 +108,7 @@ export const Content: Story = {
   },
 };
 
-export const Selection: Story = {
+export const SingleSelection: Story = {
   render: function Render() {
     const options = [
       { name: 'Bruteforce' },
@@ -119,14 +124,14 @@ export const Selection: Story = {
       { name: 'Potential Attack' },
     ];
 
-    const [attack, setAttack] = useState<SelectKey>('Bruteforce');
+    const [attack, setAttack] = useState<Selection>(new Set(['Bruteforce']));
 
     return (
       <FlexBox direction="column" gap="s">
         <Select
           items={options}
           label="Attack type"
-          selectedKey={attack}
+          selectedKeys={attack}
           style={{ inlineSize: 200 }}
           placeholder="Select an option"
           onSelectionChange={(selected) => setAttack(selected!)}
@@ -134,6 +139,73 @@ export const Selection: Story = {
           {(item) => <Select.Item key={item.name}>{item.name}</Select.Item>}
         </Select>
         <Typography>Selected: {attack}</Typography>
+      </FlexBox>
+    );
+  },
+};
+
+export const MultipleSelection: Story = {
+  render: function Render() {
+    const options = [
+      { name: 'Bruteforce' },
+      { name: 'Complex Attack' },
+      { name: 'DDoS' },
+      { name: 'DoS' },
+      { name: 'HIPS Alert' },
+      { name: 'IDS/IPS Alert' },
+      { name: 'Identity Theft' },
+      { name: 'Miscellaneous' },
+      { name: 'Network Attack' },
+      { name: 'Post Compromise' },
+      { name: 'Potential Attack' },
+    ];
+
+    const [attack, setAttack] = useState<Selection>(new Set(['Bruteforce']));
+
+    return (
+      <FlexBox direction="column" gap="s" style={{ inlineSize: 200 }}>
+        <Select
+          items={options}
+          label="Attack type"
+          selectionMode="multiple"
+          selectedKeys={attack}
+          style={{ inlineSize: 'inherit' }}
+          placeholder="Select an option"
+          onSelectionChange={(selected) => setAttack(selected!)}
+        >
+          {(item) => <Select.Item key={item.name}>{item.name}</Select.Item>}
+        </Select>
+        <Typography>Selected: {Array.from(attack).join(', ')}</Typography>
+      </FlexBox>
+    );
+  },
+};
+
+export const SelectedTagsOverflow: Story = {
+  render: function Render() {
+    return (
+      <FlexBox gap="m" direction="column">
+        <Select
+          items={options}
+          label="Attack type"
+          placeholder="Select an option"
+          selectionMode="multiple"
+          caption="selectedTagsOverflow = responsive (default)"
+          style={{ inlineSize: 220 }}
+        >
+          {(item) => <Select.Item>{item.name}</Select.Item>}
+        </Select>
+        <Select
+          items={options}
+          label="Attack type"
+          placeholder="Select an option"
+          selectionMode="multiple"
+          caption="selectedTagsOverflow = multiline"
+          style={{ inlineSize: 220 }}
+          selectedTagsOverflow="multiline"
+        >
+          {(item) => <Select.Item>{item.name}</Select.Item>}
+        </Select>
       </FlexBox>
     );
   },
@@ -238,6 +310,39 @@ export const FullWidth: Story = {
   },
 };
 
+export const ClearButton: Story = {
+  render: function Render() {
+    return (
+      <Select
+        items={options}
+        label="Attack type"
+        placeholder="Select an option"
+        defaultSelectedKeys={[1]}
+        isClearable
+      >
+        {(item) => <Select.Item>{item.name}</Select.Item>}
+      </Select>
+    );
+  },
+};
+
+export const Addons: Story = {
+  render: function Render() {
+    return (
+      <Select
+        startAddon={<IconCrosshairs16 />}
+        items={options}
+        label="Attack type"
+        style={{ inlineSize: 200 }}
+        placeholder="Select an option"
+        defaultSelectedKeys={[1]}
+      >
+        {(item) => <Select.Item>{item.name}</Select.Item>}
+      </Select>
+    );
+  },
+};
+
 export const Open: Story = {
   render: function Render() {
     const [isOpen, { toggle, set }] = useBoolean(false);
@@ -264,7 +369,7 @@ export const Open: Story = {
 export const WithIcons: Story = {
   name: 'With icons',
   render: function Render() {
-    const [selected, setSelected] = useState<SelectKey>('First');
+    const [selected, setSelected] = useState<Selection>(new Set(['First']));
 
     const options = [
       { id: 'First', icon: IconServer16 },
@@ -276,7 +381,7 @@ export const WithIcons: Story = {
 
     return (
       <Select
-        selectedKey={selected}
+        selectedKeys={selected}
         items={options}
         label="Options"
         onSelectionChange={(key) => setSelected(key!)}
@@ -297,7 +402,7 @@ export const WithIcons: Story = {
 export const WithItemDetails: Story = {
   name: 'With item details',
   render: function Render() {
-    const [selected, setSelected] = useState<SelectKey>('First');
+    const [selected, setSelected] = useState<Selection>(new Set(['First']));
 
     const options = [
       { id: 'First', caption: 'Helper text' },
@@ -309,7 +414,7 @@ export const WithItemDetails: Story = {
 
     return (
       <Select
-        selectedKey={selected}
+        selectedKeys={selected}
         items={options}
         label="Options"
         onSelectionChange={(key) => setSelected(key!)}
@@ -364,6 +469,153 @@ export const Section: Story = {
           </Select.Section>
         )}
       </Select>
+    );
+  },
+};
+
+export const StateHelper: Story = {
+  parameters: {
+    layout: 'padded',
+  },
+  render: function Render() {
+    const users = [
+      {
+        id: 1,
+        firstName: 'Emily',
+        lastName: 'Carter',
+        email: 'emily.carter@example.com',
+        role: 'Admin',
+      },
+      {
+        id: 2,
+        firstName: 'Michael',
+        lastName: 'Thompson',
+        email: 'michael.thompson@mail.com',
+        role: 'Editor',
+      },
+      {
+        id: 3,
+        firstName: 'Sophia',
+        lastName: 'Bellmont',
+        email: 'sophia.bellmont@example.org',
+        role: 'Subscriber',
+      },
+      {
+        id: 4,
+        firstName: 'Daniel',
+        lastName: 'Nguyen',
+        email: 'daniel.nguyen@mail.com',
+        role: 'User',
+      },
+      {
+        id: 5,
+        firstName: 'Olivia',
+        lastName: 'Brooks',
+        email: 'olivia.brooks@example.com',
+        role: 'Moderator',
+      },
+      {
+        id: 6,
+        firstName: 'James',
+        lastName: 'Harris',
+        email: 'james.harris@mail.org',
+        role: 'User',
+      },
+      {
+        id: 7,
+        firstName: 'Isabella',
+        lastName: 'Murphy',
+        email: 'isabella.murphy@mail.com',
+        role: 'Subscriber',
+      },
+      {
+        id: 8,
+        firstName: 'Benjamin',
+        lastName: 'Lee',
+        email: 'benjamin.lee@example.com',
+        role: 'Editor',
+      },
+      {
+        id: 9,
+        firstName: 'Ava',
+        lastName: 'Garcia',
+        email: 'ava.garcia@mail.com',
+        role: 'User',
+      },
+      {
+        id: 10,
+        firstName: 'William',
+        lastName: 'Martinez',
+        email: 'william.martinez@ex.org',
+        role: 'Admin',
+      },
+      {
+        id: 11,
+        firstName: 'Mia',
+        lastName: 'Robinson',
+        email: 'mia.robinson@mail.org',
+        role: 'Subscriber',
+      },
+      {
+        id: 12,
+        firstName: 'Alexander',
+        lastName: 'Walker',
+        email: 'alex.walker@example.com',
+        role: 'User',
+      },
+      {
+        id: 13,
+        firstName: 'Charlotte',
+        lastName: 'Scott',
+        email: 'charlotte.scott@mail.com',
+        role: 'Moderator',
+      },
+      {
+        id: 14,
+        firstName: 'Henry',
+        lastName: 'Adams',
+        email: 'henry.adams@example.org',
+        role: 'User',
+      },
+      {
+        id: 15,
+        firstName: 'Harper',
+        lastName: 'Bell',
+        email: 'harper.bell@mail.org',
+        role: 'Subscriber',
+      },
+    ];
+
+    const list = useListData({
+      initialItems: users,
+      getKey: (item) => item.id,
+      initialSelectedKeys: [3, 6, 11],
+    });
+
+    return (
+      <FlexBox direction="column" gap="m">
+        <Select
+          label="Assigned to"
+          selectionMode="multiple"
+          placeholder="Select a user"
+          style={{ minInlineSize: 200, maxInlineSize: 400, inlineSize: '100%' }}
+          items={list.items}
+          selectedKeys={list.selectedKeys}
+          onSelectionChange={(key) => list.setSelectedKeys(key)}
+        >
+          {(item) => (
+            <Select.Item textValue={item.firstName}>
+              {item.firstName} {item.lastName}
+            </Select.Item>
+          )}
+        </Select>
+        <Typography>
+          Selected emails:{' '}
+          {Array.from(list.selectedKeys)
+            .map((id) => list.getItem(id)?.email)
+            .join(', ')}
+        </Typography>
+      </FlexBox>
     );
   },
 };
