+++ packages/components/src/components/Select/__tests__/Select.test.tsx
@@ -4,10 +4,7 @@ import { screen, render } from '@testing-library/react';
 import { userEvent } from '@testing-library/user-event';
 import { describe, it, expect, vi } from 'vitest';
 
-import {
-  SelectNext as Select,
-  type SelectNextProps as SelectProps,
-} from '../index';
+import { Select, type SelectProps } from '../index';
 
 describe('SelectNext', () => {
   const baseProps: SelectProps<object> = {
