--- packages/components/src/components/SelectNext/Select.mdx
@@ -1,151 +1,0 @@
-import {
-  Meta,
-  Story,
-  Props,
-  Status,
-} from '../../../../../.storybook/components';
-
-import * as Stories from './Select.stories';
-
-<Meta of={Stories} />
-
-# SelectNext
-
-<Status variant="experimental" />
-
-A select displays a collapsible list of options and allows a user to select one or more of them.
-
-```tsx
-import { Select } from '@koobiq/react-components';
-```
-
-<Story of={Stories.Base} />
-
-## Props
-
-<Props />
-
-## Content
-
-Select accepts static and dynamic collections.The examples above show static collections,
-which can be used when the full list of options is known ahead of time.
-Dynamic collections, as shown below, can be used when the options come from an external
-data source such as an API call, or update over time.
-
-As seen below, an iterable list of options is passed to the `Select` using the `items` prop.
-Each item accepts an `key` prop, which is passed to the `onSelectionChange` handler to identify
-the selected item. Alternatively, if the item objects contain an key property, as shown
-in the example below, then this is used automatically and an `key` prop is not required.
-
-<Story of={Stories.Content} />
-
-## Single Selection
-
-Setting a selected option can be done by using the `defaultSelectedKey` or `selectedKey` prop.
-The selected key corresponds to the `key` prop of an item.
-
-<Story of={Stories.SingleSelection} />
-
-## Multiple Selection
-
-You can use the `selectionMode="multiple"` prop to allow multiple selection.
-
-<Story of={Stories.MultipleSelection} />
-
-### Selected value tags
-
-In `selectionMode="multiple"`, selected values are displayed as tags by default.
-
-To control how tags are displayed, use the `selectedTagsOverflow` prop. It supports the following values:
-
-- `responsive` (default) — tags collapse into a summary (e.g., "3 more") to fit the available width without increasing the component block size.
-- `multiline` — tags wrap to multiple lines, increasing the component height as needed.
-
-<Story of={Stories.SelectedTagsOverflow} />
-
-## Invalid
-
-The `isInvalid` prop toggles the error state. The `errorMessage` shows a message to explain the error to the user.
-
-<Story of={Stories.Invalid} />
-
-## Disabled
-
-When the select component is disabled, it cannot be interacted with.
-
-<Story of={Stories.Disabled} />
-
-## Disabled options
-
-Select supports marking items as disabled using the `disabledKeys` prop.
-Each key in this list corresponds with the `key` prop passed to the `Select.Item` component.
-
-<Story of={Stories.DisabledOptions} />
-
-## Required
-
-To make a select required, add the `isRequired` prop.
-If the field has a label, a required indicator will appear next to it.
-
-<Story of={Stories.Required} />
-
-## Full width
-
-The `fullWidth` prop will make a select fit to its parent width.
-
-<Story of={Stories.FullWidth} />
-
-## Clear Button
-
-If you pass the `isClearable` prop to the select, it will have a clear
-button which will be visible only when a value is selected.
-
-<Story of={Stories.ClearButton} />
-
-## Addons
-
-You can add an extra content using the `startAddon` and `endAddon` props.
-
-<Story of={Stories.Addons} />
-
-## Open
-
-### Default open
-
-Select isn't opened by default. The `defaultOpen` prop can be used to set the default state.
-
-### Controlled open
-
-The `isOpen` prop can be used to make the opened state controlled. The `onOpenChange` event is fired when the select's open state changes.
-
-<Story of={Stories.Open} />
-
-## Section
-
-A select component can display items grouped together in sections.
-
-<Story of={Stories.Section} />
-
-## With icons
-
-See the example below how to use [icons](?path=/docs/icons--docs) in items of the component.
-
-<Story of={Stories.WithIcons} />
-
-## With item details
-
-See the example below for using details in items of the component.
-
-<Story of={Stories.WithItemDetails} />
-
-## State Helper
-
-Use the `useListData` hook to manage items with convenient methods for adding,
-updating, and removing entries. Pass an initial set of items and a function to extract
-a unique key for each item. The returned state can then be used to render component options.
-
-```ts
-import { useListData } from '@koobiq/react-componenets';
-```
-
-<Story of={Stories.StateHelper} />
