--- packages/components/src/components/SelectNext/Select.module.css
@@ -1,80 +1,0 @@
-.base {
-  --field-input-padding-inline-start: var(--kbq-size-m);
-  --field-input-padding-inline-end: var(--kbq-size-m);
-  --field-input-padding-block-start: var(--kbq-size-xs);
-  --field-input-padding-block-end: var(--kbq-size-xs);
-  position: relative;
-  display: inline-flex;
-  justify-content: center;
-  flex-direction: column;
-  align-items: flex-start;
-}
-
-.fullWidth {
-  inline-size: 100%;
-}
-
-/* addons */
-.addon {
-  pointer-events: none;
-
-  & > :not(.chevron) {
-    pointer-events: all;
-  }
-}
-
-.chevron {
-  inline-size: var(--kbq-size-xxl);
-  block-size: var(--kbq-size-xxl);
-  padding-inline-end: var(--kbq-size-xxs);
-  margin-inline-end: calc(-1 * var(--kbq-size-s));
-  display: inline-flex;
-
-  & > svg {
-    margin: auto;
-  }
-}
-
-/* list */
-.list {
-  inline-size: 100%;
-  padding: var(--kbq-size-xxs);
-}
-
-/* popover */
-.popover {
-  border-radius: var(--kbq-size-s);
-  opacity: 0;
-  transform: translateY(-8px);
-}
-
-.popover[data-transition='entering'] {
-  opacity: 1;
-  transform: translateY(0);
-}
-
-.popover[data-transition='entered'] {
-  opacity: 1;
-  transform: translateY(0);
-}
-
-.popover[data-transition='exiting'] {
-  opacity: 0;
-  transform: translateY(-8px);
-}
-
-.popover[data-transition='exited'] {
-  opacity: 0;
-  transform: translateY(-8px);
-}
-
-.clearButton {
-  margin-inline-end: var(--kbq-size-xxs);
-}
-
-/* clear-button */
-.clearable {
-  --field-input-padding-inline-end: calc(
-    var(--field-control-end-addon-inline-size) + 28px
-  );
-}
