--- packages/components/src/components/SelectNext/index.ts
@@ -1,2 +1,0 @@
-export * from './Select';
-export * from './types';
